refactor(app): add component directory for better structure, rename TaskInput -> TaskNameInput
